Major refactoring.
