+++ src/ca/ucalgary/seng300/a3/enums/OutputMethod.java
@@ -1,4 +1,4 @@
-package enumTypes;
+package ca.ucalgary.seng300.a3.enums;
 
 public enum OutputMethod {
 	TEXT_LOG(0), DISPLAY(1), LOOPING_MESSAGE(2);
