+++ src/ca/ucalgary/seng300/a3/test/DisplayModuleTest.java
@@ -10,7 +10,7 @@ import org.junit.*;
 import org.lsmr.vending.*;
 import org.lsmr.vending.hardware.*;
 
-import ca.ucalgary.seng300.a3.*;
+import ca.ucalgary.seng300.a3.core.*;
 
 /**
  * Software Engineering 300 - Group Assignment 2
