+++ src/ca/ucalgary/seng300/a3/test/TransactionModuleTest.java
@@ -10,7 +10,8 @@ import org.junit.*;
 import org.lsmr.vending.*;
 import org.lsmr.vending.hardware.*;
 
-import ca.ucalgary.seng300.a3.*;
+import ca.ucalgary.seng300.a3.core.*;
+import ca.ucalgary.seng300.a3.exceptions.InsufficientFundsException;
 
 /**
  * Software Engineering 300 - Group Assignment 2
